requirements.txt
# requirements.txt
streamlit
mysql-connector-python